requirements.txt
langchain[openai]
python-dotenv
langgraph
langchain_mcp_adapters
streamlit
google.genai
langchain-chroma
langchain_community
bs4
langchain_openai
langchain-ollama
ollama
